remove big-integer dependency

+++ src/io-ts/Integer.ts
@@ -1,9 +1,9 @@
 import { unsafeCoerce } from 'fp-ts/lib/function'
 import { Type } from 'io-ts'
 import { Integer as IntegerNewtype } from '../Integer'
-import { BigInteger } from './BigInteger'
+import { bigint } from './bigint'
 
 /**
  * @since 0.1.2
  */
-export const Integer: Type<IntegerNewtype, string, unknown> = unsafeCoerce(BigInteger)
+export const Integer: Type<IntegerNewtype, string, unknown> = unsafeCoerce(bigint)
